APIBasicRAG_chatbot/deploy2/requirements.txt
requests
streamlit
openpyxl
pymongo 
python-dotenv